+++ _import/tycho-execution/src/encoding/models.rs
@@ -1,3 +1,4 @@
+use alloy_primitives::PrimitiveSignature as Signature;
 use hex;
 use num_bigint::BigUint;
 use serde::{Deserialize, Serialize};
@@ -7,8 +8,7 @@ use tycho_common::{
 };
 
 use crate::encoding::{
-    errors::EncodingError,
-    serde_primitives::{biguint_string, biguint_string_option},
+    errors::EncodingError, evm::approvals::permit2::PermitSingle, serde_primitives::biguint_string,
 };
 
 /// Represents a solution containing details describing an order, and  instructions for filling
@@ -30,15 +30,9 @@ pub struct Solution {
     /// supported.
     #[serde(default)]
     pub exact_out: bool,
-    /// If set, it will be applied to expected_amount
-    pub slippage: Option<f64>,
-    /// Expected amount of the bought token (exact in) or sold token (exact out).
-    #[serde(with = "biguint_string_option")]
-    pub expected_amount: Option<BigUint>,
     /// Minimum amount to be checked for the solution to be valid.
-    /// If not set, the check will not be performed.
-    #[serde(with = "biguint_string_option")]
-    pub checked_amount: Option<BigUint>,
+    #[serde(with = "biguint_string")]
+    pub checked_amount: BigUint,
     /// List of swaps to fulfill the solution.
     pub swaps: Vec<Swap>,
     /// If set, the corresponding native action will be executed.
@@ -94,6 +88,25 @@ pub struct Transaction {
     pub to: Bytes,
     pub value: BigUint,
     pub data: Vec<u8>,
+}
+
+/// Represents a solution that has been encoded for execution.
+///
+/// # Fields
+/// * `swaps`: Encoded swaps to be executed.
+/// * `interacting_with`: Address of the contract to be called.
+/// * `selector`: The selector of the function to be called.
+/// * `n_tokens`: Number of tokens in the swap.
+/// * `permit`: Optional permit for the swap (if permit2 is enabled).
+/// * `signature`: Optional signature for the swap (if permit2 is enabled).
+#[derive(Clone, Debug)]
+pub struct EncodedSolution {
+    pub swaps: Vec<u8>,
+    pub interacting_with: Bytes,
+    pub selector: String,
+    pub n_tokens: usize,
+    pub permit: Option<PermitSingle>,
+    pub signature: Option<Signature>,
 }
 
 /// Represents necessary attributes for encoding an order.